Generate hanzi struct at compile time
Generate the hanzi struct in Hanzi.Map at compile time. This offers the
compiler more optimization opportunities (since the struct is constant
instead of generated at runtime), at the cost of making the Hanzi.Map
module dependent on the Hanzi module. This is not a problem when using
hanyutils as a dependency, but painful when developing the hanzi module.