hanzi: Remove characters with strange pinyin
Since we can't parse them, we can't deal with them in a useful manner.
When the internal representation of the Pinyin module is changed to
internal/final/tone we can't represent these characters at all anymore,
because we have to parse them.

When the

Strangeness:

- n with tone marker
- Standalone o
- Strange standalone finals:
  - ng
  - n
  - m
  - wong

I don't think it would be useful to adapt the parsers to include these
because we'd parse a lot of strange sequences.
Not sure if they are "allowed" in the "official" Pinyin?

The standalone o is not super strange, people actually use the character 哦

TODO:
- Modify the converter to generate this properly.
- Don't remove commonly used characters